Create Regiser form in auth page
